refactor: GlobalExceptionHandler 구현
